Re-send second CB-message during update if it appears to have been missed or gets requested again
Certain (unofficial) devices (namely the AskSin-OTA-Bootloader ones) are having issues when the second CB-message is sent too fast so they will miss it and the update will fail. This commit retries the transmission once if the response isn't received (that could mean the message arrived too early and so it got missed) or if the bootloader asks for it by sending another update request.